requirements.txt
###### Requirements without Version Specifiers ######
setuptools
biopython
numpy
wheel
tabulate